implementing binary and linear search, update on main.py
