+++ README.md
@@ -1,1 +1,29 @@
-# Computer-Science
+Advanced Algorithm module at University College Birmingham
+
+Collaborators
+Luis Henrique Treve Nunes (GitHub, LinkedIn)
+Dawood Madarshahian (GitHub, LinkedIn)
+
+Overview
+This project was developed as part of the Advanced Algorithm module at University College Birmingham. It includes various types of algorithms, including quick sort, merge sort, binary search, linear search, GRAPH algorithms, and machine learning algorithms.
+
+Installation
+To install the program, you need a Python interpreter and Git Bash terminal to run the main.py script. To download the source code, simply clone the repository using the following command: git clone https://github.com/Gooz95/Computer-Science.git
+
+Usage
+To use the program, navigate to the project directory and run the main.py script using the following command: python main.py
+
+Dependencies
+This project does not have any external dependencies.
+
+Contribution
+We welcome contributions from the community. To contribute to the project, please fork the repository, make your changes, and submit a pull request. We will review your changes and merge them if they meet our quality standards.
+
+Bug Reports and Feature Requests
+If you encounter any bugs or have any feature requests, please contact us by email or through the issue tracker on GitHub.
+
+License
+This project is licensed under the MIT License.
+
+Contact Us
+If you have any questions or issues, please feel free to contact us by email or through our LinkedIn profiles.
